Move view classes to view package.
