+++ src/main/java/uk/gov/register/presentation/view/RegisterProofView.java
@@ -1,4 +1,4 @@
-package uk.gov.register.thymeleaf;
+package uk.gov.register.presentation.view;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import io.dropwizard.views.View;
